Fix Opencensus tracing. (#117)
When passing a lambda to runWithRetries, the grpc context is lost,
causing spans in OpenCensus to become disconnected. This PR fixes the
issue by calling the wrap function in the grpc library, which wraps a
callable with the appropriate context so that it's propagated all the
way throughout.

The OpenCensus test is also changed to include the use of the query and
commit endpoints for more thorough testing.