Update Java v1.8.x support in README.md (#146)
Fixes #144.

Not adding a test because of following reasons:

Can't add the dependency in dgraph4j as it will enforce it for all java versions, which in not what we want.
Can't add it in DgraphJavaSample, as people will be looking at it irrespective of java version they use. So, I think its better for it to stay as is. Just the documentation update should be enough to guide people who face problems with Java 1.8.x